refactor: rename to update states from sources
