+++ src/shared/loom-state/update-state-from-sources.tsx
